requirements.txt
pydantic==1.8.1
uvicorn==0.13.4
fastapi==0.63.0
websockets==8.1
PyJWT==2.0.1
python-decouple==3.4
aiofiles==0.6.0
python-multipart==0.0.5
jinja2==2.11.2
email-validator==1.1.2
psycopg2-binary==2.8.6
websockets==8.1